Test is_already_processed and refactor
